+++ assets_rebalance_backend/src/Domain/FinAssetsGroup.cs
@@ -6,7 +6,7 @@ namespace assets_rebalance_backend.src.Domain;
 public class FinAssetsGroup
 {
     public required string Name { get; set; }
-    public required int Score { get; set; }
+    public required float Score { get; set; }
     public IEnumerable<FinAsset> Children { get; set; } = [];
     public decimal CurrentAmount => Children.Sum(x => x.CurrentAmount);
     public decimal ScorePercent => (decimal)Score / 100;
